web/requirements.txt
Django==1.11
gunicorn==19.7.0
psycopg2==2.7
redis==2.10.3
djangorestframework==3.10.3